services/api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
ajsonrpc==1.2.0
annotated-types==0.7.0
anyio==4.9.0
atomicwrites==1.4.1
attrs==25.3.0
blinker==1.9.0
bottle==0.13.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
coverage==7.8.0
deprecation==2.1.0
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.6.0
gotrue==2.12.0
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.3
packaging==24.2
paho-mqtt==1.6.1
pgvector==0.1.8
pillow==11.2.1
platformio==6.1.18
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
psycopg2-binary==2.9.10
py==1.11.0
pydantic==2.11.3
pydantic_core==2.33.1
pyelftools==0.32
PyJWT==2.10.1
pyserial==3.5
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
realtime==2.4.2
requests==2.31.0
semantic-version==2.10.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tabulate==0.9.0
toml==0.10.2
twilio==8.9.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0